// src/app/api/model/getVehicles200ResponseInner.ts
/**
 * Vehicle API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface GetVehicles200ResponseInner {
    id?: string;
    name?: string;
    manufacturer?: string;
    model?: string;
    fuel?: string;
    type?: string;
    vin?: string;
    color?: string | null;
    mileage?: number | null;
}